fix(exercises): accept non-integer durationMs in practice session schema
Every exercise completion was silently failing PracticeSessionInputSchema
validation: useExercisePhaseTimer accumulates durationMs from
requestAnimationFrame timestamps, which are sub-millisecond floats, but the
schema required z.number().int(). This meant savePracticeSession returned
{success:false} on every single completion since the placeholder was
replaced with real persistence — found while verifying the first end-to-end
exercise-tracking flow (start → complete → persist → see updated progress).

Verified fix with a real signed-in user completing all six Eye Foundation
exercises end-to-end: status persists across refreshes, replaying an
already-completed exercise updates completed_at without creating a
duplicate row, and the landing page reflects each completion immediately.

+++ src/lib/exercises/types.ts
@@ -8,7 +8,8 @@ export const PracticeSessionInputSchema = z
   .object({
     labId: LabIdSchema,
     exerciseId: z.string().min(1),
-    durationMs: z.number().int().positive(),
+    // rAF timestamps are sub-millisecond floats — not an integer by nature.
+    durationMs: z.number().positive(),
     completed: z.boolean(),
   })
   .strict()
